agentforce/requirements.txt
# Python dependencies for Agentforce agent management
pyyaml>=6.0.1
